updated course and institute delete method

+++ api/admin/institute.php
@@ -3,10 +3,12 @@
 namespace api\admin;
 
 use api\Controller;
+use model\StudentModel;
 use model\InstituteModel;
 use middleware\AuthMiddleware;
 
 require_once(__DIR__ . "/../../model/InstituteModel.php");
+require_once(__DIR__ . "/../../model/StudentModel.php");
 require_once(__DIR__ . "/../../middleware/AuthMiddleware.php");
 require_once(__DIR__ . "/../Controller.php");
 
@@ -156,6 +158,14 @@ class Institute extends Controller
 			response(404, false, ["message" => "Institute not found!"]);
 			exit;
 		}
+		//verify if student is enrolled in the institute
+		$studentByMajor = StudentModel::find($results["slug"], "institute");
+
+		if ($studentByMajor) {
+			response(400, false, ["message" => "Students are enrolled in this institute"]);
+			exit;
+		}
+
 
 		if (InstituteModel::delete($id, "id")) {
 			response(200, true, ["message" => "Delete successful"]);
